fix: move render table/block functions into separate util file
