+++ apps/google-docs/src/locations/Page/components/review/utils/documentOutlineUtils.ts
@@ -1,4 +1,5 @@
-import type { EntryBlockGraph, NormalizedDocumentContentBlock, WorkflowContentType } from '@types';
+import type { DocSegment } from '../utils/buildDocument';
+import type { EntryBlockGraph, WorkflowContentType } from '@types';
 
 export interface ListMarker {
   marker: string;
@@ -43,9 +44,10 @@ const FIELD_TYPE_LABELS: Record<string, string> = {
 export const getFieldTypeLabel = (fieldType: string): string =>
   FIELD_TYPE_LABELS[fieldType] ?? fieldType;
 
-export function buildListMarkers(
-  blocks: NormalizedDocumentContentBlock[]
-): Record<string, ListMarker> {
+export function buildListMarkers(segments: DocSegment[]): Record<string, ListMarker> {
+  const blocks = segments
+    .filter((seg): seg is Extract<DocSegment, { kind: 'block' }> => seg.kind === 'block')
+    .map((seg) => seg.block);
   const listMarkers: Record<string, ListMarker> = {};
   const orderedCounts = new Map<number, number>();
 
@@ -85,6 +87,7 @@ export function buildListMarkers(
   return listMarkers;
 }
 
+// TODO: replace this with the overview section component once it is compatible with entryBlockGraph
 export function buildOverviewEntries(
   entries: EntryBlockGraph['entries'],
   contentTypes: WorkflowContentType[]
